Update No.1, testing additional conversion and bug in PrefixCollection significant digits

+++ UnitConverter/PrefixCollection.cs
@@ -2,7 +2,7 @@
 
 namespace UnitConverterLib
 {
-    public class PrefixCollection : Dictionary<string, int> 
+    public class PrefixCollection : Dictionary<string, double> 
     {
         
     }
